Split tests into two parts that run in parallel in CI
