+++ tests/phpunit/RestoreTests/AbsRestoreTestPart1.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\ProjectRestore\Tests\RestoreTests;
+
+use Keboola\ProjectRestore\Tests\AbstractRestoreTestPart1;
+
+class AbsRestoreTestPart1 extends AbstractRestoreTestPart1
+{
+    use AbsRestoreTestTrait;
+}
