+++ tests/phpunit/RestoreTests/AbsRestoreTestPart2.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\ProjectRestore\Tests\RestoreTests;
+
+use Keboola\ProjectRestore\Tests\AbstractRestoreTestPart2;
+
+class AbsRestoreTestPart2 extends AbstractRestoreTestPart2
+{
+    use AbsRestoreTestTrait;
+}
