+++ tests/phpunit/RestoreTests/S3RestoreTestPart2.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Keboola\ProjectRestore\Tests\RestoreTests;
+
+use Keboola\ProjectRestore\Tests\AbstractRestoreTestPart2;
+
+class S3RestoreTestPart2 extends AbstractRestoreTestPart2
+{
+    use S3RestoreTestTrait;
+}
